requirements.txt
langchain==0.3.27
langchain-groq==0.3.8
langchain-core==0.3.83
langgraph==1.0.1
chromadb==1.4.1
sentence-transformers==5.2.0
streamlit==1.41.1
python-dotenv==1.2.1
